requirements.txt
beautifulsoup4==4.10.0
bokeh==2.4.2
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
filelock==3.4.2
gdown==4.2.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.22.1
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
tornado==6.1
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.8
